Added a volume script updated readme
The volume script is a bit more easy to handle. Some of the client names are
a bit annoying, e.g. "ALSA plug-in [mpv]", which is made easier to change with
this script. Some may also have several occurances, e.g. "Firefox Developer
Edition" and "Firefox" can appear, and the script changes them both.